reserve_app/requirements.txt
django
django-allauth
django-bootstrap5
pillow
qrcode